Fix bugs in function public.list_users()
This function didn't work as expected as it was a security definer
function. The calls to pg_has_role were emitted by the owner of the
function, here pgfactory, and didn't give the expected result.

Also, this function made the assumption that there's only one super
user called 'postgres'. Changed to take account on all super user
from a PostgreSQL cluster.