Refactor/breaks grapher.js to ajax-refresh graphs
Warning: this patch probably breaks and spit acid on the graphs/edit
page! It only focus on the graphs/show page!
This patch allows to refresh graphs on zoom in/out with fresh data
fetched from the server. A large cleanup has been done on the way which
probably break the edit page and the preview.