processes: add extension pr_grapher_wh_nagios.
Dependancy between pr_grapher and wh_nagios are now located in a specific
extension. This extension use schema "pr_grapher", and depends on
pgfactory_core, pr_grapher and wh_nagios.

Functions that used to use wh_nagios tables are relocated (and renamed) into
this extension.

The list_graph remaining function is now only available to superuser, and
only lists graphs, without any information on a warehouse.

CAREFUL: with this commit, the graph UI is broken.