[UI] list services binded to graphs on host page insteads of all graphs
From the hostname page, we now list all services of the host instead of
all the graphs usings its services. We now link to a new page showing
all the graphs using the given service.

It makes the hot page much smaller and useful. Next patch might add some
links on this page to jump on one specifi graph only.