[nagios_dispatcher] Fix infinite loop on perfdata insert error
On error when inserting perfdata, the algorithm close the connexion
and re-run the current block that is supposed to open a new connexion.
However, even after calling $dbh->disconnect, $dhb is still true, thus
the block does not open a new connexion, leading to another error and
the infinite loop.